+++ packages/orq-rc/src/models/operations/updateidentity.ts
@@ -167,7 +167,7 @@ export const UpdateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-14T10:23:22.152Z",
+    "2026-05-14T12:03:57.631Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {