+++ packages/orq-rc/RELEASES.md
@@ -9189,3 +9189,13 @@ Based on:
 - [typescript v4.10.0-rc.4] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.4 - packages/orq-rc
+
+## 2026-05-14 12:07:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.0 (2.884.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.5] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.5 - packages/orq-rc